Moved the vdb load to openvdb_asset.cpp, loading dll at VolumeLoader constructor
